refactor: update file paths for stylesheets
