--- src/App.scss
@@ -1,5 +1,0 @@
-@use './utils/variables' as *;
-
-.welcome {
-  color: $primary-color;
-}
