+++ src/App.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { RouterProvider } from "react-router-dom";
 import router from "./routes";
-import "./App.scss";
+import "./styles/App.scss";
 
 function App() {
   return <RouterProvider router={router} />;
