--- src/index.scss
@@ -1,13 +1,0 @@
-@use './utils/variables' as *;
-
-body {
-  margin: 0;
-  display: grid;
-  place-items: center;
-  min-width: 320px;
-  min-height: 100vh;
-  font-family: $primary-font;
-  background-color: $white;
-}
-
-
