--- src/utils/_variables.scss
@@ -1,25 +1,0 @@
-// Opticode color palette
-@import url('https://fonts.googleapis.com/css2?family=JetBrains+Mono:ital,wght@0,100..800;1,100..800&family=Lato:ital,wght@0,100;0,300;0,400;0,700;0,900;1,100;1,300;1,400;1,700;1,900&family=Marcellus&family=Open+Sans:ital,wght@0,300..800;1,300..800&family=Poppins:ital,wght@0,100;0,200;0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,100;1,200;1,300;1,400;1,500;1,600;1,700;1,800;1,900&family=Roboto:ital,wght@0,100;0,300;0,400;0,500;0,700;0,900;1,100;1,300;1,400;1,500;1,700;1,900&display=swap');
-
-// Fonts
-$primary-font: Lato, system-ui, Avenir, Helvetica, Arial, sans-serif;
-$secondary-font: JetBrains, "Roboto", sans-serif;
-
-// Primary Colors
-$primary-color: #1E3A8A;
-$secondary-color: #2D9CDB;
-$tertiary-color: #A0E7E5;
-
-// Neutral Colors
-$white: #F3F4F6;
-$light-gray: #CBD6EB;
-$gray: #6B7280;
-$black: #07090E;
-
-// Accent Colors
-$yellow: #F2C94C;
-$red: #EB5757;
-$green: #27AE60;
-
-
-
